renamed roomUser
